Add cross platform and testing support to hscpp.

+++ examples/hscpp-example-utils/include/hscpp-example-utils/IMemoryManager.h
@@ -5,7 +5,7 @@
 class IMemoryManager
 {
 public:
-    virtual uint8_t* GetMemory(size_t id) = 0;
+    virtual uint8_t* GetMemory(uint64_t id) = 0;
 };
